Update .gitignore and save entered passwords to a JSON file
This commit updates the .gitignore file to ignore .json files instead of .txt files. It also modifies the main.py file to save entered passwords to a JSON file instead of a text file. The passwords are saved in the new_data dictionary and are written to the data.json file. If the file doesn't exist, a new one is created. If the file already exists, it is updated with the new_data.